requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Pillow==10.0.0
numpy==1.24.0
scikit-learn==1.3.0
skin-tone-classifier==1.0.0
gunicorn==21.2.0
